requirements.txt
aiofiles==23.2.1
aiogram==2.25.1
aiohttp==3.8.5
aiomysql==0.2.0
aiosignal==1.3.1
anyio==4.0.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Deprecated==1.2.14
dotenv-python==0.0.1
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-resources==6.0.1
limits==3.6.0
magic-filter==1.0.11
multidict==6.0.4
packaging==23.1
pycparser==2.21
pylast==5.2.0
PyMySQL==1.1.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
sniffio==1.3.0
typing_extensions==4.7.1
urllib3==2.0.4
wrapt==1.15.0
yandex-music==2.1.1
yarl==1.9.2
